refactor: moving interfaces into types file
